requirements.txt
tensorflow>=2.5.0
numpy
opencv-python
scikit-learn
fastapi 
uvicorn
pillow
python-multipart